Fix for unique key not found on vnode
Animation always work on last component added or removed because in the list of inferno-transition-group. It could not get unique key of components from the list. This patch will fix that problem and getting keys from vnode works with vnode.key instead of vnode.attributes.key. This may be because of the inferno vnode update in new version.